refactor: split up `main`
- command related stuff moved into `command`
- client instantiation moved into `client`
- the `Error` type moved into `error`
- printing utilities moved into `fmt`
