+++ src/main.rs
@@ -1,35 +1,14 @@
-use std::{
-    collections::HashMap,
-    convert::{TryFrom, TryInto},
-    env,
-    ffi::OsStr,
-    fmt, fs, io, iter,
-    path::{Path, PathBuf},
-    process,
-    str::FromStr,
-    time::Duration,
-};
+mod client;
+mod command;
+mod error;
+mod fmt;
 
-use clap::{CommandFactory, Parser};
-use cloudformatious::{
-    change_set::ChangeSet, status_reason::StatusReasonDetail, ApplyStackError, ApplyStackInput,
-    Capability, CloudFormatious, DeleteStackError, DeleteStackInput, Parameter, StackEvent,
-    StackFailure, StackStatus, StackWarning, StatusSentiment, TemplateSource,
-};
-use colored::{ColoredString, Colorize};
-use futures_util::{Stream, StreamExt};
-use rusoto_cloudformation::{CloudFormationClient, Tag};
-use rusoto_core::{HttpClient, Region};
-use rusoto_credential::{
-    AutoRefreshingProvider, ChainProvider, ProfileProvider, ProvideAwsCredentials,
-};
+use std::process;
 
-const MISSING_REGION: &str = "Unable to determine AWS region.
-You can set it in your profile, assign `AWS_REGION`, or supply `--region`.";
-const AWS_CLOUDFORMATION_STACK: &str = "AWS::CloudFormation::Stack";
-const SHORT_UPDATE_COMPLETE_CLEANUP_IN_PROGRESS: &str = "UPDATE_CLEANUP_IN_PROGRESS";
-const SHORT_UPDATE_ROLLBACK_COMPLETE_CLEANUP_IN_PROGRESS: &str = "ROLLBACK_CLEANUP_IN_PROGRESS";
-const NO_REASON: &str = "No reason";
+use clap::Parser;
+use rusoto_core::Region;
+
+use self::{client::get_client, error::Error};
 
 /// A CloudFormation CLI that won't make you cry.
 ///
@@ -40,448 +19,19 @@ const NO_REASON: &str = "No reason";
 #[derive(Parser, Debug)]
 #[clap(name = "cloudformatious")]
 struct Args {
-    /// Disable informational output to STDERR.
-    #[clap(long)]
-    quiet: bool,
-
     /// The region to use. Overrides config/env settings.
     #[clap(long, env = "AWS_REGION")]
     region: Option<Region>,
 
     #[clap(subcommand)]
-    command: Command,
-}
-
-#[derive(Parser, Debug)]
-enum Command {
-    Completions(CompletionsArgs),
-    ApplyStack(ApplyStackArgs),
-    DeleteStack(DeleteStackArgs),
-}
-
-/// Write a shell completion script to STDOUT.
-#[derive(Parser, Debug)]
-struct CompletionsArgs {
-    /// The shell to generate completions for.
-    ///
-    /// Determined from $SHELL if left unset.
-    #[clap(value_enum)]
-    shell: Option<clap_complete::Shell>,
-}
-
-fn default_shell() -> Option<clap_complete::Shell> {
-    let shell = env::var("SHELL").ok().map(PathBuf::from);
-    shell
-        .as_deref()
-        .and_then(Path::file_name)
-        .and_then(OsStr::to_str)
-        .and_then(|shell_name| shell_name.parse().ok())
-}
-
-/// Apply a CloudFormation template.
-///
-/// This performs an update or create operation for a target stack. It's not an error for there
-/// to be no changes. The command runs until the stack settles.
-///
-/// # Output
-///
-/// Stack events are printed to STDERR as the operation proceeds, unless disabled with `--quiet`.
-///
-/// If the stack operation succeeds and there are no resource errors, then the stack's outputs
-/// are printed to STDOUT as JSON.
-///
-/// If the stack operation succeeds and there *are* resource errors, then details of the errors
-/// are printed to STDERR and the stack's outputs are printed to STDOUT as JSON.
-///
-/// If the stack operation fails, then details of the error(s) are printed to STDERR.
-///
-/// # Exit code
-///
-/// If the stack operation succeeds and there are no resource errors, then the CLI will exit
-/// successfully with code 0.
-///
-/// If the operation succeeds but there *are* resource errors, then the exit code is 3.
-///
-/// If the operation fails because the stack settled in an error state, then exit code is 4.
-///
-/// If the operation fails for any other reason, then the exit code is 1.
-#[derive(Parser, Debug)]
-struct ApplyStackArgs {
-    /// Capabilities to explicitly acknowledge.
-    #[clap(long)]
-    capabilities: Vec<CapabilityArg>,
-
-    /// A unique identifier for this `apply_stack` operation.
-    #[clap(long)]
-    client_request_token: Option<String>,
-
-    /// The Simple Notification Service (SNS) topic ARNs to publish stack related events.
-    #[clap(long)]
-    notification_arns: Vec<String>,
-
-    /// A list of input parameters for the stack.
-    ///
-    /// Parameters should be supplied as `key=value` strings.
-    #[clap(long)]
-    parameters: Vec<ParameterArg>,
-
-    /// The template resource types that you have permissions to work with for this `apply_stack`
-    /// operation, such as `AWS::EC2::Instance`, `AWS::EC2::*`, or `Custom::MyCustomInstance`.
-    #[clap(long)]
-    resource_types: Vec<String>,
-
-    /// The Amazon Resource Name (ARN) of an AWS Identity And Access Management (IAM) role that AWS
-    /// CloudFormation assumes to apply the stack.
-    #[clap(long)]
-    role_arn: Option<String>,
-
-    /// The name that is associated with the stack.
-    ///
-    /// If this isn't set explicitly then the file name of the `template_path` is used as the stack
-    /// name. E.g. if `template_path` is `deployment/cloudformation/my-stack.yaml` then the default
-    /// stack name would be `my-stack`.
-    #[clap(long)]
-    stack_name: Option<String>,
-
-    /// Key-value pairs to associate with this stack.
-    ///
-    /// Tags should be supplied either as `key=value` strings and/or as a JSON object (e.g.
-    /// `{"key1": "value1", "key2": "value2"}). JSON is tried first.
-    #[clap(long)]
-    tags: Vec<TagArg>,
-
-    /// Path to the template to be applied.
-    template_path: PathBuf,
-}
-
-impl TryFrom<ApplyStackArgs> for ApplyStackInput {
-    type Error = Error;
-    fn try_from(args: ApplyStackArgs) -> Result<Self, Self::Error> {
-        let template_path = args.template_path;
-
-        Ok(ApplyStackInput {
-            capabilities: args.capabilities.into_iter().map(Into::into).collect(),
-            client_request_token: args.client_request_token,
-            notification_arns: args.notification_arns,
-            parameters: args.parameters.into_iter().map(Into::into).collect(),
-            resource_types: if args.resource_types.is_empty() {
-                None
-            } else {
-                Some(args.resource_types)
-            },
-            role_arn: args.role_arn,
-            stack_name: args.stack_name.unwrap_or_else(|| {
-                template_path
-                    .file_stem()
-                    .unwrap_or_else(|| OsStr::new(""))
-                    .to_string_lossy()
-                    .to_string()
-            }),
-            tags: args.tags.into_iter().flatten().collect(),
-            template_source: TemplateSource::inline(fs::read_to_string(&template_path).map_err(
-                |error| {
-                    Error::other(format!(
-                        "Invalid template path {:?}: {}",
-                        template_path, error
-                    ))
-                },
-            )?),
-        })
-    }
-}
-
-/// Newtype for parsing capabilities.
-// TODO: use impl Deserialize upstream and use `Capability` directly.
-#[derive(Debug)]
-struct CapabilityArg(Capability);
-
-impl FromStr for CapabilityArg {
-    type Err = InvalidCapability;
-    fn from_str(capability: &str) -> Result<Self, Self::Err> {
-        let capability = match capability {
-            "CAPABILITY_IAM" => Capability::Iam,
-            "CAPABILITY_NAMED_IAM" => Capability::NamedIam,
-            "CAPABILITY_AUTO_EXPAND" => Capability::AutoExpand,
-            _ => return Err(InvalidCapability(capability.to_string())),
-        };
-        Ok(Self(capability))
-    }
-}
-
-impl From<CapabilityArg> for Capability {
-    fn from(arg: CapabilityArg) -> Self {
-        arg.0
-    }
-}
-
-#[derive(Debug)]
-struct InvalidCapability(String);
-
-impl fmt::Display for InvalidCapability {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "invalid capability `{}`, should be one of `{}`, `{}`, or `{}`",
-            self.0,
-            Capability::Iam,
-            Capability::NamedIam,
-            Capability::AutoExpand
-        )
-    }
-}
-
-impl std::error::Error for InvalidCapability {}
-
-/// Newtype for parsing parameters.
-#[derive(Debug)]
-struct ParameterArg(Parameter);
-
-impl FromStr for ParameterArg {
-    type Err = InvalidParameter;
-    fn from_str(parameter: &str) -> Result<Self, Self::Err> {
-        let kv: Vec<_> = parameter.splitn(2, '=').collect();
-        let [key, value]: [_; 2] = kv
-            .try_into()
-            .map_err(|_| InvalidParameter(parameter.to_string()))?;
-        Ok(Self(Parameter {
-            key: key.to_string(),
-            value: value.to_string(),
-        }))
-    }
-}
-
-impl From<ParameterArg> for Parameter {
-    fn from(arg: ParameterArg) -> Self {
-        arg.0
-    }
-}
-
-#[derive(Debug)]
-struct InvalidParameter(String);
-
-impl fmt::Display for InvalidParameter {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "invalid parameter `{}`, must be in the form `key=value`",
-            self.0
-        )
-    }
-}
-
-impl std::error::Error for InvalidParameter {}
-
-/// Newtype for parsing tags.
-#[derive(Debug)]
-enum TagArg {
-    KeyValue(Tag),
-    Json(Vec<Tag>),
-}
-
-impl FromStr for TagArg {
-    type Err = InvalidTag;
-    fn from_str(tag: &str) -> Result<Self, Self::Err> {
-        // First try to parse as JSON
-        let tags: Result<HashMap<String, String>, _> = serde_json::from_str(tag);
-        if let Ok(tags) = tags {
-            return Ok(TagArg::Json(
-                tags.into_iter()
-                    .map(|(key, value)| Tag { key, value })
-                    .collect(),
-            ));
-        }
-
-        let kv: Vec<_> = tag.splitn(2, '=').collect();
-        let [key, value]: [_; 2] = kv.try_into().map_err(|_| InvalidTag(tag.to_string()))?;
-        Ok(Self::KeyValue(Tag {
-            key: key.to_string(),
-            value: value.to_string(),
-        }))
-    }
-}
-
-impl IntoIterator for TagArg {
-    type Item = Tag;
-    type IntoIter = std::vec::IntoIter<Self::Item>;
-    fn into_iter(self) -> Self::IntoIter {
-        match self {
-            Self::KeyValue(tag) => vec![tag].into_iter(),
-            Self::Json(tags) => tags.into_iter(),
-        }
-    }
-}
-
-#[derive(Debug)]
-struct InvalidTag(String);
-
-impl fmt::Display for InvalidTag {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "invalid tag `{}`, must be in the form `key=value` or a JSON object",
-            self.0
-        )
-    }
-}
-
-impl std::error::Error for InvalidTag {}
-
-/// Delete a CloudFormation stack.
-///
-/// # Output
-///
-/// Stack events are printed to STDERR as the operation proceeds, unless disable with `--quiet`.
-///
-/// If the stack is deleted successfully and there are no resource errors, or if no stack
-/// existed in the first place, a confirmation message is printed to STDERR.
-///
-/// If the stack is deleted successfully and there *are* resource errors, then details of the
-/// errors are printed to STDERR.
-///
-/// If the stack deletion fails, then details of the error(s) are printed to STDERR.
-///
-/// # Exit code
-///
-/// If the stack is deleted successfully and there are no resource errors, or if no stack
-/// existed in the first place, then the CLI will exit successfully with code 0.
-///
-/// If the stack is deleted successfully but there *are* resource errors, then the exit code is
-/// 3.
-///
-/// If the stack deletion fails because the stack settled in an error state, then exit code is
-/// 4.
-///
-/// If the deletion fails for any other reason, then the exit code is 1.
-#[derive(Parser, Debug)]
-struct DeleteStackArgs {
-    /// A unique identifier for this `delete_stack` operation.
-    #[clap(long)]
-    client_request_token: Option<String>,
-
-    /// For stacks in the `DELETE_FAILED` state, a list of resource logical IDs that are associated
-    /// with the resources you want to retain. During deletion, AWS CloudFormation deletes the stack
-    /// but does not delete the retained resources.
-    #[clap(long)]
-    retain_resources: Vec<String>,
-
-    /// The Amazon Resource Name (ARN) of an AWS Identity And Access Management (IAM) role that AWS
-    /// CloudFormation assumes to delete the stack.
-    #[clap(long)]
-    role_arn: Option<String>,
-
-    /// The name of the stack to delete.
-    #[clap(long)]
-    stack_name: Option<String>,
-
-    /// The path to the template whose associated stack will be deleted.
-    ///
-    /// The stack to delete is determined from the file name of `template_path`. E.g. if
-    /// `template_path` is `deployment/cloudformation/my-stack.yaml` then the default `my-stack`
-    /// will be deleted.
-    #[clap(required_unless_present = "stack-name")]
-    template_path: Option<PathBuf>,
-}
-
-impl TryFrom<DeleteStackArgs> for DeleteStackInput {
-    type Error = Error;
-    fn try_from(args: DeleteStackArgs) -> Result<Self, Self::Error> {
-        let template_path = args.template_path;
-        Ok(DeleteStackInput {
-            client_request_token: args.client_request_token,
-            retain_resources: if args.retain_resources.is_empty() {
-                None
-            } else {
-                Some(args.retain_resources)
-            },
-            role_arn: args.role_arn,
-            stack_name: args.stack_name.unwrap_or_else(|| {
-                template_path
-                    .expect("bug: DeleteStackArgs without stack_name or template_path")
-                    .file_stem()
-                    .unwrap_or_else(|| OsStr::new(""))
-                    .to_string_lossy()
-                    .to_string()
-            }),
-        })
-    }
-}
-
-#[derive(Debug)]
-enum Error {
-    Warning(StackWarning),
-    Failure(StackFailure),
-    Other(Box<dyn std::error::Error>),
-}
-
-impl Error {
-    fn other<E: Into<Box<dyn std::error::Error>>>(error: E) -> Self {
-        Self::Other(error.into())
-    }
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Warning(warning) => warning.fmt(f),
-            Self::Failure(failure) => {
-                writeln!(f, "Failed to apply stack {}:\n", failure.stack_id.bold())?;
-
-                let status = failure.stack_status.to_string();
-                writeln!(f, "   {} {}", "Status:".bold(), status.red())?;
-                writeln!(f, "   {} {}", "Reason:".bold(), failure.stack_status_reason)?;
-
-                if let Some(hint) = failure.stack_status_reason().detail().and_then(get_hint) {
-                    writeln!(f, "   {:<7} {}", "Hint:".bold(), hint)?;
-                }
-
-                if !failure.resource_events.is_empty() {
-                    writeln!(f, "\nWhat went wrong? The following resource errors occurred during the operation:")?;
-                    for (index, (resource_status, event_details)) in
-                        failure.resource_events.iter().enumerate()
-                    {
-                        let resource = event_details.logical_resource_id();
-                        let type_ = event_details.resource_type();
-                        let status = resource_status.to_string();
-                        let reason = event_details
-                            .resource_status_reason()
-                            .inner()
-                            .unwrap_or(NO_REASON);
-                        writeln!(f, "\n{}. {} {}", index + 1, "Resource:".bold(), resource)?;
-                        writeln!(f, "   {:<9} {}", "Type:".bold(), type_)?;
-                        writeln!(f, "   {:<9} {}", "Status:".bold(), status.red())?;
-                        writeln!(f, "   {:<9} {}", "Reason:".bold(), reason)?;
-
-                        if let Some(hint) = event_details
-                            .resource_status_reason()
-                            .detail()
-                            .and_then(get_hint)
-                        {
-                            writeln!(f, "   {:<9} {}", "Hint:".bold(), hint)?;
-                        }
-                    }
-                }
-
-                Ok(())
-            }
-            Self::Other(error) => error.fmt(f),
-        }
-    }
-}
-
-impl std::error::Error for Error {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        match self {
-            Self::Warning(_) | Self::Failure(_) => None,
-            Self::Other(error) => Some(error.as_ref()),
-        }
-    }
+    command: command::Command,
 }
 
 #[tokio::main]
 async fn main() {
     let args = Args::parse();
 
-    if let Err(error) = try_main(args).await {
+    if let Err(error) = command::main(args.region, args.command).await {
         eprintln!("{}", error);
         process::exit(match error {
             Error::Warning(_) => 3,
@@ -489,267 +39,4 @@ async fn main() {
             Error::Other(_) => 1,
         });
     }
-}
-
-async fn try_main(args: Args) -> Result<(), Error> {
-    let client = get_client(
-        args.region
-            .map(Ok)
-            .or_else(get_region)
-            .ok_or_else(|| Error::other(MISSING_REGION))?
-            .map_err(Error::other)?,
-    )
-    .await
-    .map_err(Error::other)?;
-
-    match args.command {
-        Command::Completions(cmd_args) => {
-            let mut cmd = Args::command();
-            let cmd_name = cmd.get_name().to_string();
-            if let Some(shell) = cmd_args.shell.or_else(default_shell) {
-                clap_complete::generate(shell, &mut cmd, cmd_name, &mut io::stdout());
-            } else {
-                let shell_env = env::var("SHELL");
-                return Err(Error::other(format!(
-                    "unable to determine current shell from $SHELL ({})",
-                    shell_env.as_deref().unwrap_or("not set"),
-                )));
-            }
-        }
-        Command::ApplyStack(cmd_args) => {
-            let mut apply = client.apply_stack(cmd_args.try_into()?);
-
-            let change_set = apply.change_set().await.map_err(Error::other)?;
-            let sizing = Sizing::new_for_change_set(&change_set);
-
-            if !args.quiet {
-                print_events(&sizing, apply.events()).await;
-            }
-
-            let output = apply.await.map_err(|error| match error {
-                ApplyStackError::Warning { warning, .. } => Error::Warning(warning),
-                ApplyStackError::Failure(failure) => Error::Failure(failure),
-                ApplyStackError::CloudFormationApi(_) => Error::other(error),
-                ApplyStackError::CreateChangeSetFailed { .. } => Error::other(error),
-            })?;
-
-            let outputs_json: serde_json::Value = output
-                .outputs
-                .into_iter()
-                .map(|output| (output.key, output.value.into()))
-                .collect::<serde_json::Map<_, _>>()
-                .into();
-            println!(
-                "{}",
-                serde_json::to_string_pretty(&outputs_json).expect("oh no")
-            );
-        }
-        Command::DeleteStack(cmd_args) => {
-            let mut delete = client.delete_stack(cmd_args.try_into()?);
-            let sizing = Sizing::default();
-
-            if !args.quiet {
-                print_events(&sizing, delete.events()).await;
-            }
-
-            delete.await.map_err(|error| match error {
-                DeleteStackError::Warning(warning) => Error::Warning(warning),
-                DeleteStackError::Failure(failure) => Error::Failure(failure),
-                DeleteStackError::CloudFormationApi(_) => Error::other(error),
-            })?;
-        }
-    }
-
-    Ok(())
-}
-
-fn get_region() -> Option<Result<Region, Box<dyn std::error::Error>>> {
-    // rusoto_cloudformation::Region::default implements a similar algorithm but falls back to
-    // us-east-1, which we don't want.
-    match env::var("AWS_DEFAULT_REGION").or_else(|_| env::var("AWS_REGION")) {
-        Ok(region) => Some(region.parse().map_err(Into::into)),
-        Err(_) => match ProfileProvider::region() {
-            Ok(Some(region)) => Some(region.parse().map_err(Into::into)),
-            Ok(None) => None,
-            Err(error) => Some(Err(error.into())),
-        },
-    }
-}
-
-async fn get_client(region: Region) -> Result<CloudFormationClient, Box<dyn std::error::Error>> {
-    let client = HttpClient::new()?;
-
-    let mut credentials = ChainProvider::new();
-    credentials.set_timeout(Duration::from_secs(1));
-
-    let credentials =
-        AutoRefreshingProvider::new(aws_sso_flow::ChainProvider::new().push(credentials).push(
-            aws_sso_flow::SsoFlow::builder().verification_prompt(|url| async move {
-                if atty::is(atty::Stream::Stdin) && atty::is(atty::Stream::Stdout) {
-                    eprintln!("Using SSO an profile – go to {url} to authenticate");
-                    Ok(())
-                } else {
-                    Err(NonInteractiveSsoError)
-                }
-            }),
-        ))?;
-
-    // Proactively fetch credentials so we get earlier errors.
-    credentials.credentials().await?;
-
-    Ok(CloudFormationClient::new_with(client, credentials, region))
-}
-
-#[derive(Debug)]
-struct NonInteractiveSsoError;
-
-impl fmt::Display for NonInteractiveSsoError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "can't complete SSO authentication in a non-interactive context"
-        )
-    }
-}
-
-impl std::error::Error for NonInteractiveSsoError {}
-
-struct Sizing {
-    resource_status: usize,
-    logical_resource_id: usize,
-    resource_type: usize,
-}
-
-impl Sizing {
-    fn new_for_change_set(change_set: &ChangeSet) -> Self {
-        let default = Self::default();
-        Self {
-            resource_status: default.resource_status,
-            logical_resource_id: change_set
-                .changes
-                .iter()
-                .map(|change| change.logical_resource_id.len())
-                .chain(iter::once(change_set.stack_name.len()))
-                .max()
-                .unwrap(), // we insert the stack name so unwrap is fine
-            resource_type: change_set
-                .changes
-                .iter()
-                .map(|change| change.resource_type.len())
-                .chain(iter::once(default.resource_type))
-                .max()
-                .unwrap(), // we insert the default so unwrap is fine
-        }
-    }
-}
-
-impl Default for Sizing {
-    fn default() -> Self {
-        Self {
-            resource_status: SHORT_UPDATE_ROLLBACK_COMPLETE_CLEANUP_IN_PROGRESS.len(),
-            logical_resource_id: 0,
-            resource_type: AWS_CLOUDFORMATION_STACK.len(),
-        }
-    }
-}
-
-async fn print_events(sizing: &Sizing, mut events: impl Stream<Item = StackEvent> + Unpin) {
-    while let Some(event) = events.next().await {
-        eprintln!(
-            "{:?} {:resource_status_size$} {:logical_resource_id_size$} {:resource_type_size$} {}",
-            event.timestamp(),
-            colorize_status(&event),
-            event.logical_resource_id(),
-            event.resource_type(),
-            event.resource_status_reason().unwrap_or("").bright_black(),
-            resource_status_size = sizing.resource_status,
-            logical_resource_id_size = sizing.logical_resource_id,
-            resource_type_size = sizing.resource_type,
-        );
-    }
-    eprintln!();
-}
-
-fn colorize_status(event: &StackEvent) -> ColoredString {
-    let status = match event {
-        StackEvent::Resource {
-            resource_status, ..
-        } => resource_status.to_string(),
-        // Shorten the most verbose statuses for better formatting
-        StackEvent::Stack {
-            resource_status, ..
-        } => match resource_status {
-            StackStatus::UpdateCompleteCleanupInProgress => {
-                SHORT_UPDATE_COMPLETE_CLEANUP_IN_PROGRESS.to_string()
-            }
-            StackStatus::UpdateRollbackCompleteCleanupInProgress => {
-                SHORT_UPDATE_ROLLBACK_COMPLETE_CLEANUP_IN_PROGRESS.to_string()
-            }
-            _ => resource_status.to_string(),
-        },
-    };
-    match event.resource_status().sentiment() {
-        StatusSentiment::Positive => status.green(),
-        StatusSentiment::Neutral => status.yellow(),
-        StatusSentiment::Negative => status.red(),
-    }
-}
-
-fn get_hint(detail: StatusReasonDetail) -> Option<String> {
-    match detail {
-        StatusReasonDetail::CreationCancelled => Some("See preceding resource errors".to_string()),
-        StatusReasonDetail::MissingPermission(detail) => Some(format!(
-            "Give {} the {} permission",
-            detail
-                .principal
-                .map(Colorize::bold)
-                .unwrap_or_else(|| "yourself".normal()),
-            detail.permission.bold()
-        )),
-        StatusReasonDetail::ResourceErrors(detail) => Some(format!(
-            "See resource error(s) for {}",
-            display_list(detail.logical_resource_ids().map(|id| id.bold()))
-        )),
-        _ => None,
-    }
-}
-
-fn display_list<I, T>(iter: I) -> impl fmt::Display
-where
-    I: IntoIterator<Item = T>,
-    T: fmt::Display,
-{
-    use fmt::Write;
-
-    let mut output = String::new();
-
-    let mut iter = iter.into_iter().peekable();
-    let mut seen = 0;
-    while let Some(item) = iter.next() {
-        if seen > 0 {
-            let next = iter.peek();
-            if seen > 1 || next.is_some() {
-                output.push(',');
-            }
-            if next.is_none() {
-                output.push_str(" and");
-            }
-            output.push(' ');
-        }
-        write!(&mut output, "{}", item).unwrap();
-        seen += 1;
-    }
-
-    output
-}
-
-#[test]
-fn test_display_list() {
-    assert_eq!(&display_list(&[1]).to_string(), "1");
-
-    assert_eq!(&display_list(&[1, 2]).to_string(), "1 and 2");
-
-    assert_eq!(&display_list(&[1, 2, 3]).to_string(), "1, 2, and 3");
-
-    assert_eq!(&display_list(&[1, 2, 3, 4]).to_string(), "1, 2, 3, and 4");
 }
